move schedulers to schedulers namespace
