+++ tt-media-server/cpp_server/tests/ipc_scheduler_smoke_test.cpp
@@ -25,8 +25,8 @@
 #include "domain/sampling_params.hpp"
 #include "domain/sequence.hpp"
 #include "ipc/boost_ipc_task_queue.hpp"
-#include "runners/llm_runner/prefill_first_scheduler.hpp"
-#include "runners/llm_runner/scheduler.hpp"
+#include "runners/llm_runner/schedulers/prefill_first_scheduler.hpp"
+#include "runners/llm_runner/schedulers/scheduler.hpp"
 
 using Sequence = tt::domain::Sequence;
 using SamplingParams = tt::domain::SamplingParams;
@@ -37,7 +37,7 @@ static constexpr size_t MAX_NUM_MSGS = 64;
 static constexpr size_t MAX_MSG_SIZE = 4096;
 
 int main() {
-  using namespace tt::runners::llm_engine;
+  using namespace tt::runners::schedulers;
   using Config = tt::config::LLMConfig;
 
   // Clean up any leftover queue from a previous failed run.
